chore(route-recognizer): move parseQueryString from kernel to route-recognizer and remove kernel dep
