+++ packages/route-recognizer/package.json
@@ -34,9 +34,7 @@
   "publishConfig": {
     "access": "public"
   },
-  "dependencies": {
-    "@aurelia/kernel": "0.6.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "tslib": "^1.10.0",
     "typescript": "^3.7.4"
